Refactor UI system into modular handlers
