+++ src/core/ui/panelHandlers.ts
@@ -1,30 +1,19 @@
 import * as mc from '@minecraft/server';
-import { ModalFormData, ActionFormData, ActionFormResponse, ModalFormResponse } from '@minecraft/server-ui';
+import { ActionFormResponse, ModalFormResponse } from '@minecraft/server-ui';
 
-import * as bountyManager from '../bountyManager.js';
-import { getConfig, updateMultipleConfig, resetConfigSection } from '../configManager.js';
-import { getShopConfig, getXrayConfig, saveXrayConfig, XrayConfig } from '../configurations.js';
-import { floatingTextManager } from '../floatingTextManager.js';
-import * as helpfulLinksManager from '../helpfulLinksManager.js';
-import { items as allItems } from '../itemsConfig.default.js';
-import * as kitAdminManager from '../kitAdminManager.js';
-import { errorLog } from '../logger.js';
-import { getValueFromPath, setValueByPath } from '../objectUtils.js';
-import { getPlayer, PlayerData, loadPlayerData } from '../playerDataManager.js';
-import * as rankManager from '../rankManager.js';
-import * as reportManager from '../reportManager.js';
-import * as rulesManager from '../rulesManager.js';
-import * as shopAdminManager from '../shopAdminManager.js';
-import * as shopManager from '../shopManager.js';
+import { getPlayer } from '../playerDataManager.js';
 import { showPanel } from '../uiManager.js';
-import * as utils from '../utils.js';
-import { MonitoredOreType } from '../xrayConfig.default.js';
 
 import { handleUIAction } from './actions.js';
-import { showConfirmationDialog } from './components.js';
+import { handleAdminPanel } from './handlers/adminHandlers.js';
+import { handleConfigPanel } from './handlers/configHandlers.js';
+import { handleGameplayPanel } from './handlers/gameplayHandlers.js';
+import { handleMiscPanel } from './handlers/miscHandlers.js';
+import { handlePlayerPanel } from './handlers/playerHandlers.js';
+import { handleShopPanel } from './handlers/shopHandlers.js';
+import { handleTeamPanel } from './handlers/teamHandlers.js';
 import { getMenuItems, getVisiblePlayerActionItems } from './panelBuilder.js';
-import { configPanelSchema, UIContext, panelDefinitions, PanelItem } from './panelRegistry.js';
-import { itemsPerPage, getPaginatedItems, configHandlers, getVisibleConfigSystems } from './uiUtils.js';
+import { panelDefinitions, PanelItem, UIContext } from './panelRegistry.js';
 
 export async function handleFormResponse(
     player: mc.Player,
@@ -32,7 +21,6 @@ export async function handleFormResponse(
     response: ActionFormResponse | ModalFormResponse,
     context: UIContext
 ) {
-    const { canceled } = response;
     const pData = getPlayer(player.id);
     if (!pData) {
         return;
@@ -40,7 +28,6 @@ export async function handleFormResponse(
 
     // Helper properties with type guards implicitly handled by usage context
     const selection = (response as ActionFormResponse).selection;
-    const formValues = (response as ModalFormResponse).formValues;
 
     // Generic handler for registry-defined panels
     const panelDef = panelDefinitions[panelId];
@@ -82,2441 +69,51 @@ export async function handleFormResponse(
         }
     }
 
-    if (panelId === 'floatingTextListPanel') {
-        if (selection === 0) {
-            // Back
-            return showPanel(player, 'mainPanel', context);
-        }
-        if (selection === 1) {
-            // Create New
-            return showPanel(player, 'floatingTextCreatePanel', context);
-        }
-        const texts = floatingTextManager.getAllTexts();
-        if (typeof selection === 'number') {
-            const selectedText = texts[selection - 2];
-            if (selectedText) {
-                return showPanel(player, 'floatingTextActionPanel', { ...context, id: selectedText.id });
-            }
-        }
-        return;
+    // Delegate to specialized handlers
+    if (panelId.startsWith('shop')) {
+        return handleShopPanel(player, panelId, response, context);
     }
 
-    if (panelId === 'floatingTextActionPanel') {
-        if (typeof selection === 'number') {
-            const { id } = context;
-            try {
-                switch (selection) {
-                    case 0: // Edit
-                        return showPanel(player, 'floatingTextEditPanel', context);
-                    case 1: // Respawn
-                        await floatingTextManager.respawnText(id);
-                        player.sendMessage(`§2Respawned floating text: ${id}`);
-                        break;
-                    case 2: // Despawn
-                        await floatingTextManager.despawnText(id);
-                        player.sendMessage(`§2Despawned floating text: ${id}`);
-                        break;
-                    case 3: // Delete
-                        await floatingTextManager.deleteText(player, id);
-                        break; // The deleteText function sends its own success message.
-                    case 4: // Back
-                        break;
-                }
-            } catch (error) {
-                errorLog(`[UIManager] Error in floatingTextActionPanel for ID '${id}':`, error);
-                player.sendMessage('§4An error occurred. Please check the logs.');
-            }
-            // Always refresh the list panel, even on error or 'Back'
-            return showPanel(player, 'floatingTextListPanel', context);
-        }
-        return;
+    if (panelId.startsWith('team')) {
+        return handleTeamPanel(player, panelId, response, context);
     }
 
-    if (panelId === 'reportListPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'mainPanel', context);
+    if (panelId.startsWith('player')) {
+        return handlePlayerPanel(player, panelId, response, context);
+    }
 
-        const reports = reportManager
-            .getAllReports()
-            .filter((r) => r.status === 'open' || r.status === 'assigned')
-            .sort((a, b) => a.timestamp - b.timestamp);
+    if (panelId.startsWith('floating') || panelId.startsWith('report')) {
+        return handleAdminPanel(player, panelId, response, context);
+    }
 
-        const hasPrev = page > 1;
-        let buttonIndex = selection - 1;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-
-        const paginatedReports = getPaginatedItems(reports, page);
-        if (buttonIndex >= 0 && buttonIndex < paginatedReports.length) {
-            const report = paginatedReports[buttonIndex];
-            return showPanel(player, 'reportActionsPanel', { ...context, targetReport: report });
-        }
-        buttonIndex -= paginatedReports.length;
-
-        const totalPages = Math.ceil(reports.length / itemsPerPage);
-        if (page < totalPages && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
+    if (
+        panelId.startsWith('rules') ||
+        panelId.startsWith('helpfulLink') ||
+        panelId.startsWith('addRule') ||
+        panelId.startsWith('addHelpful') ||
+        panelId.startsWith('ruleAction')
+    ) {
+        return handleMiscPanel(player, panelId, response, context);
     }
 
     if (panelId === 'bountyListPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'mainPanel', context);
-
-        const allBounties = Array.from(bountyManager.getAllBounties().entries())
-            .map(([id, val]) => ({ ...val, targetId: id }))
-            .sort((a, b) => b.amount - a.amount);
-
-        const hasPrev = page > 1;
-        let buttonIndex = selection - 1;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-
-        const paginatedBounties = getPaginatedItems(allBounties, page);
-        if (buttonIndex >= 0 && buttonIndex < paginatedBounties.length) {
-            const bounty = paginatedBounties[buttonIndex];
-            return showPanel(player, 'playerActionsPanel', {
-                ...context,
-                targetPlayerId: bounty.targetId,
-                targetPlayerName: bounty.name,
-                fromPanel: 'bountyListPanel'
-            });
-        }
-        buttonIndex -= paginatedBounties.length;
-
-        const totalPages = Math.ceil(allBounties.length / itemsPerPage);
-        if (page < totalPages && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
+        return handleGameplayPanel(player, panelId, response, context);
     }
 
-    if (panelId === 'playerListPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'infoPanel', context);
-        if (selection === 1)
-            return showPanel(player, 'playerSearchPanel', { ...context, fromPanel: 'playerListPanel' });
-
-        const onlinePlayers = Array.from(mc.world.getAllPlayers()).sort((a, b) => a.name.localeCompare(b.name));
-
-        const hasPrev = page > 1;
-        let buttonIndex = selection - 2;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-
-        const paginatedPlayers = getPaginatedItems(onlinePlayers, page);
-        if (buttonIndex >= 0 && buttonIndex < paginatedPlayers.length) {
-            const target = paginatedPlayers[buttonIndex];
-            return showPanel(player, 'playerActionsPanel', {
-                ...context,
-                targetPlayerId: target.id,
-                targetPlayerName: target.name,
-                fromPanel: 'playerListPanel'
-            });
-        }
-        buttonIndex -= paginatedPlayers.length;
-
-        const totalPages = Math.ceil(onlinePlayers.length / itemsPerPage);
-        if (page < totalPages && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
+    // Config catch-all (includes config_, kit_, rank_, command_, economy_, xray_)
+    if (
+        panelId.startsWith('config') ||
+        panelId.startsWith('kit') ||
+        panelId.startsWith('rank') ||
+        panelId.startsWith('command') ||
+        panelId.startsWith('economy') ||
+        panelId.startsWith('mob') ||
+        panelId.startsWith('addMob') ||
+        panelId.startsWith('editMob') ||
+        panelId.startsWith('xray') ||
+        panelId.startsWith('addXray') ||
+        panelId.startsWith('editXray')
+    ) {
+        return handleConfigPanel(player, panelId, response, context);
     }
-
-    if (panelId === 'playerManagementPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'adminPanel', context);
-        if (selection === 1)
-            return showPanel(player, 'playerSearchPanel', { ...context, fromPanel: 'playerManagementPanel' });
-
-        const { getAllPlayerNameIdMap } = await import('../playerDataManager.js');
-        const allPlayersMap = getAllPlayerNameIdMap();
-        const playerEntries = Array.from(allPlayersMap.entries()).sort((a, b) => a[0].localeCompare(b[0]));
-
-        const hasPrev = page > 1;
-        let buttonIndex = selection - 2;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-
-        const paginatedEntries = getPaginatedItems(playerEntries, page);
-        if (buttonIndex >= 0 && buttonIndex < paginatedEntries.length) {
-            const [name, id] = paginatedEntries[buttonIndex];
-            const targetData = loadPlayerData(id);
-            const properName = targetData ? targetData.name : name;
-            return showPanel(player, 'playerActionsPanel', {
-                ...context,
-                targetPlayerId: id,
-                targetPlayerName: properName,
-                fromPanel: 'playerManagementPanel'
-            });
-        }
-        buttonIndex -= paginatedEntries.length;
-
-        const totalPages = Math.ceil(playerEntries.length / itemsPerPage);
-        if (page < totalPages && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
-    }
-
-    if (panelId === 'playerSearchPanel') {
-        if (canceled) {
-            const from = context.fromPanel || 'mainPanel';
-            return showPanel(player, from, context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [searchName] = values as string[];
-        if (!searchName) {
-            player.sendMessage('§4Search name is required.');
-            return showPanel(player, panelId, context);
-        }
-
-        const { getPlayerIdByName, getAllPlayerNameIdMap } = await import('../playerDataManager.js');
-        let targetId = getPlayerIdByName(searchName);
-        let targetName = searchName;
-
-        if (!targetId) {
-            const allPlayers = getAllPlayerNameIdMap();
-            const searchLower = searchName.toLowerCase();
-            for (const [name, id] of allPlayers.entries()) {
-                if (name.includes(searchLower)) {
-                    targetId = id;
-                    const targetData = loadPlayerData(id);
-                    targetName = targetData ? targetData.name : name;
-                    break;
-                }
-            }
-        }
-
-        if (targetId) {
-            return showPanel(player, 'playerActionsPanel', {
-                ...context,
-                targetPlayerId: targetId,
-                targetPlayerName: targetName
-            });
-        } else {
-            player.sendMessage(`§4Player '${searchName}' not found.`);
-            return showPanel(player, panelId, context);
-        }
-    }
-
-    if (panelId === 'kitManagementPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'configCategoryPanel', context);
-
-        const config = getConfig();
-        if (selection === 1) {
-            const newStatus = !config.kits.enabled;
-            updateMultipleConfig({ 'kits.enabled': newStatus });
-            player.sendMessage(`§2Kit system ${newStatus ? 'enabled' : 'disabled'}.`);
-            return showPanel(player, panelId, context);
-        }
-
-        if (selection === 2) {
-            const form = new ModalFormData().title('Create Kit').textField('Kit Name', 'Enter unique name');
-            const res = await utils.uiWait(player, form);
-            if (!res.canceled && (res as ModalFormResponse).formValues) {
-                const name = (res as ModalFormResponse).formValues![0] as string;
-                if (name) {
-                    kitAdminManager.createKit(name);
-                    player.sendMessage('§2Kit created.');
-                }
-            }
-            return showPanel(player, panelId, context);
-        }
-
-        const kitNames = Object.keys(kitAdminManager.getAllKits());
-        const paginatedKits = getPaginatedItems(kitNames, page);
-        let buttonIndex = selection - 3;
-
-        if (buttonIndex >= 0 && buttonIndex < paginatedKits.length) {
-            const kitName = paginatedKits[buttonIndex];
-            return showPanel(player, `kitActionMenu_${kitName}`, context);
-        }
-        buttonIndex -= paginatedKits.length;
-
-        const totalPages = Math.ceil(kitNames.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-        if (hasNext) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
-    }
-
-    if (panelId.startsWith('kitActionMenu_')) {
-        const kitName = panelId.replace('kitActionMenu_', '');
-        if (selection === 0) return showPanel(player, `kitSettingsPanel_${kitName}`, context);
-        if (selection === 1) return showPanel(player, `kitItemsPanel_${kitName}`, context);
-        if (selection === 2) {
-            kitAdminManager.deleteKit(kitName);
-            player.sendMessage('§2Kit deleted.');
-            return showPanel(player, 'kitManagementPanel', context);
-        }
-        if (selection === 3) return showPanel(player, 'kitManagementPanel', context);
-        return;
-    }
-
-    if (panelId.startsWith('kitSettingsPanel_')) {
-        const kitName = panelId.replace('kitSettingsPanel_', '');
-        if (canceled) return showPanel(player, `kitActionMenu_${kitName}`, context);
-
-        const values = (response as ModalFormResponse).formValues;
-        if (values) {
-            const [enabled, name, desc, icon, cooldownStr, permStr, priceStr] = values as [
-                boolean,
-                string,
-                string,
-                string,
-                string,
-                string,
-                string
-            ];
-            const cooldown = parseInt(cooldownStr) || 0;
-            const perm = parseInt(permStr) || 1024;
-            const price = parseInt(priceStr) || 0;
-
-            kitAdminManager.updateKitSettings(kitName, {
-                enabled,
-                description: desc,
-                icon,
-                cooldownSeconds: cooldown,
-                permissionLevel: perm,
-                price
-            });
-
-            if (name !== kitName && name) {
-                const res = kitAdminManager.renameKit(kitName, name);
-                player.sendMessage(res.message);
-                if (res.success) {
-                    return showPanel(player, `kitActionMenu_${name}`, context);
-                }
-            } else {
-                player.sendMessage('§2Kit settings updated.');
-            }
-        }
-        return showPanel(player, `kitActionMenu_${kitName}`, context);
-    }
-
-    if (panelId.startsWith('kitItemsPanel_')) {
-        const kitName = panelId.replace('kitItemsPanel_', '');
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) {
-            // Add Item
-            const form = new ModalFormData()
-                .title('Add Item')
-                .textField('Item ID', 'minecraft:stone')
-                .textField('Amount', '1', { defaultValue: '1' });
-            const res = await utils.uiWait(player, form);
-            if (!res.canceled && (res as ModalFormResponse).formValues) {
-                const [typeId, amountStr] = (res as ModalFormResponse).formValues as [string, string];
-                const amount = parseInt(amountStr);
-                if (typeId && !isNaN(amount)) {
-                    const { addItemToKit } = await import('../kitItemsManager.js');
-                    const result = addItemToKit(kitName, { typeId, amount });
-                    player.sendMessage(result.message);
-                }
-            }
-            return showPanel(player, panelId, context);
-        }
-
-        const kit = kitAdminManager.getAllKits()[kitName];
-        if (!kit) return showPanel(player, 'kitManagementPanel', context);
-
-        const paginatedItems = getPaginatedItems(kit.items, page);
-        let buttonIndex = selection - 1;
-
-        if (buttonIndex >= 0 && buttonIndex < paginatedItems.length) {
-            // Edit/Delete Item
-            const itemIndex = (page - 1) * itemsPerPage + buttonIndex;
-            const form = new ActionFormData()
-                .title('Manage Item')
-                .button('Delete Item', 'textures/ui/trash')
-                .button('Cancel', 'textures/ui/cancel');
-            const res = await utils.uiWait(player, form);
-            if (!res.canceled && (res as ActionFormResponse).selection === 0) {
-                const { removeItemFromKit } = await import('../kitItemsManager.js');
-                const result = removeItemFromKit(kitName, itemIndex);
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, panelId, context);
-        }
-        buttonIndex -= paginatedItems.length;
-
-        // Pagination
-        const totalPages = Math.ceil(kit.items.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-
-        if (hasNext) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page + 1 });
-            buttonIndex--;
-        }
-
-        // Back (Last button)
-        if (buttonIndex === 0) {
-            return showPanel(player, `kitActionMenu_${kitName}`, context);
-        }
-        return showPanel(player, panelId, context);
-    }
-
-    if (panelId === 'commandSystemPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'configCategoryPanel', context);
-
-        const config = getConfig();
-        const commandSettings = config.commandSettings || {};
-        const allCommands = Object.keys(commandSettings)
-            .filter((c) => !c.startsWith('_'))
-            .sort();
-
-        const paginatedCommands = getPaginatedItems(allCommands, page);
-        let buttonIndex = selection - 1;
-
-        if (buttonIndex >= 0 && buttonIndex < paginatedCommands.length) {
-            const cmd = paginatedCommands[buttonIndex];
-            return showPanel(player, 'commandSettingsPanel', { ...context, commandName: cmd });
-        }
-        buttonIndex -= paginatedCommands.length;
-
-        const totalPages = Math.ceil(allCommands.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-        if (hasNext) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
-    }
-
-    if (panelId === 'commandSettingsPanel') {
-        if (canceled) return showPanel(player, 'commandSystemPanel', context);
-        const values = (response as ModalFormResponse).formValues;
-        const { commandName } = context;
-        if (values && commandName) {
-            const [enabled, permLevelStr] = values as [boolean, string];
-            const permLevel = parseInt(permLevelStr);
-            if (!isNaN(permLevel)) {
-                updateMultipleConfig({
-                    [`commandSettings.${commandName}.enabled`]: enabled,
-                    [`commandSettings.${commandName}.permissionLevel`]: permLevel
-                });
-                player.sendMessage(`§2Updated settings for ${commandName}.`);
-            }
-        }
-        return showPanel(player, 'commandSystemPanel', context);
-    }
-
-    if (panelId === 'rankManagementPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'configCategoryPanel', context);
-        if (selection === 1) return showPanel(player, 'rankSettingsPanel', context);
-        if (selection === 2) return showPanel(player, 'addRankPanel', context);
-
-        const allRanks = rankManager.getAllRanks().sort((a, b) => a.permissionLevel - b.permissionLevel);
-        const paginatedRanks = getPaginatedItems(allRanks, page);
-        let buttonIndex = selection - 3;
-
-        if (buttonIndex >= 0 && buttonIndex < paginatedRanks.length) {
-            const rank = paginatedRanks[buttonIndex];
-            return showPanel(player, `rankActionMenu_${rank.id}`, context);
-        }
-        buttonIndex -= paginatedRanks.length;
-
-        const totalPages = Math.ceil(allRanks.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-        if (hasNext) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
-    }
-
-    if (panelId === 'addRankPanel') {
-        if (canceled) return showPanel(player, 'rankManagementPanel', context);
-        const values = (response as ModalFormResponse).formValues;
-        if (values) {
-            const [name, id, permStr, nameColor, chatColor, prefix] = values as string[];
-            const perm = parseInt(permStr);
-            if (name && id && !isNaN(perm)) {
-                const rankDb = await import('../rankDb.js');
-                const result = rankDb.addRank({
-                    id,
-                    name,
-                    permissionLevel: perm,
-                    chatFormatting: {
-                        nameColor: nameColor || '§8',
-                        messageColor: chatColor || '§r',
-                        prefixText: prefix || ''
-                    },
-                    conditions: [{ type: 'hasTag', value: id }],
-                    locked: false
-                });
-                player.sendMessage(result.message);
-                if (result.success) rankManager.reloadRanks();
-            }
-        }
-        return showPanel(player, 'rankManagementPanel', context);
-    }
-
-    if (panelId.startsWith('rankActionMenu_')) {
-        const rankId = panelId.replace('rankActionMenu_', '');
-        if (selection === 0) return showPanel(player, 'editRankPanel', { ...context, rankId });
-        if (selection === 1) {
-            // Delete
-            const rankDb = await import('../rankDb.js');
-            const result = rankDb.deleteRank(rankId);
-            player.sendMessage(result.message);
-            if (result.success) rankManager.reloadRanks();
-            return showPanel(player, 'rankManagementPanel', context);
-        }
-        if (selection === 2) return showPanel(player, 'rankManagementPanel', context);
-        return;
-    }
-
-    if (panelId === 'editRankPanel') {
-        if (canceled) return showPanel(player, 'rankManagementPanel', context);
-        const values = (response as ModalFormResponse).formValues;
-        const { rankId } = context;
-        if (values && rankId) {
-            const [name, newId, permStr, nameColor, chatColor, prefix, nametag] = values as string[];
-            const perm = parseInt(permStr);
-            if (!isNaN(perm)) {
-                const rankDb = await import('../rankDb.js');
-                const result = rankDb.updateRank(rankId, {
-                    id: newId,
-                    name,
-                    permissionLevel: perm,
-                    chatFormatting: { nameColor: nameColor, messageColor: chatColor, prefixText: prefix },
-                    nametagPrefix: nametag
-                });
-                player.sendMessage(result.message);
-                if (result.success) rankManager.reloadRanks();
-            }
-        }
-        return showPanel(player, 'rankManagementPanel', context);
-    }
-
-    if (panelId === 'rankSettingsPanel') {
-        if (canceled) return showPanel(player, 'rankManagementPanel', context);
-        const values = (response as ModalFormResponse).formValues;
-        if (values) {
-            const styleIndex = values[0] as number;
-            const styles = ['above', 'before', 'after', 'under'];
-            const style = styles[styleIndex];
-            updateMultipleConfig({ 'ranks.nameTagStyle': style });
-            player.sendMessage('§2Rank settings updated.');
-        }
-        return showPanel(player, 'rankManagementPanel', context);
-    }
-
-    if (panelId === 'teamMainPanel') {
-        const { getTeamByPlayer } = await import('../teamManager.js');
-
-        const team = getTeamByPlayer(player.id);
-
-        if (selection === 0) {
-            return showPanel(player, 'mainPanel', context);
-        }
-
-        if (team) {
-            const isOwner = team.ownerId === player.id;
-
-            const isAdmin = team.admins.includes(player.id);
-            const isOwnerOrAdmin = isOwner || isAdmin;
-
-            let btnIndex = 1;
-
-            if (selection === btnIndex) {
-                // Members
-                return showPanel(player, 'teamMembersPanel', context);
-            }
-            btnIndex++;
-
-            if (isOwnerOrAdmin) {
-                if (selection === btnIndex) {
-                    // Manage
-                    return showPanel(player, 'teamManagePanel', context);
-                }
-                btnIndex++;
-            }
-
-            if (selection === btnIndex) {
-                // Settings
-                return showPanel(player, 'teamSettingsPanel', context);
-            }
-            btnIndex++;
-
-            if (selection === btnIndex) {
-                // Leave
-                if (isOwner) {
-                    player.sendMessage(
-                        '§4Owners cannot leave their team. You must transfer ownership or delete the team.'
-                    );
-                } else {
-                    showConfirmationDialog(player, {
-                        title: 'Leave Team',
-                        body: 'Are you sure you want to leave the team?',
-                        confirmButtonText: '§4Yes, Leave',
-                        cancelButtonText: 'No',
-                        onConfirm: async () => {
-                            const { kickMember } = await import('../teamManager.js');
-                            // Kick self
-                            kickMember(team.id, player.id);
-                            player.sendMessage('§2You have left the team.');
-                            showPanel(player, 'teamMainPanel', context);
-                        },
-                        onCancel: () => showPanel(player, 'teamMainPanel', context)
-                    });
-                }
-                return;
-            }
-        } else {
-            // No Team
-            if (selection === 1) {
-                // Create
-                return showPanel(player, 'teamCreatePanel', context);
-            }
-            if (selection === 2) {
-                // Join
-                return showPanel(player, 'teamJoinPanel', context);
-            }
-        }
-        return;
-    }
-
-    if (panelId === 'teamCreatePanel') {
-        if (canceled) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [name] = values as string[];
-        if (!name) {
-            player.sendMessage('§4Team name is required.');
-            return showPanel(player, panelId, context);
-        }
-
-        const { createTeam } = await import('../teamManager.js');
-        const result = createTeam(player, name);
-        player.sendMessage(result.message || '§4Unknown error.');
-        return showPanel(player, 'teamMainPanel', context);
-    }
-
-    if (panelId === 'teamJoinPanel') {
-        if (selection === 0) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-        if (selection === 1) {
-            return showPanel(player, 'teamInvitesPanel', context);
-        }
-        if (selection === 2) {
-            return showPanel(player, 'teamSearchPanel', context);
-        }
-        if (selection === 3) {
-            return showPanel(player, 'teamBrowserPanel', context);
-        }
-        return;
-    }
-
-    if (panelId === 'teamSearchPanel') {
-        if (canceled) {
-            return showPanel(player, 'teamJoinPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [idStr] = values as string[];
-        const teamId = parseInt(idStr);
-        if (isNaN(teamId)) {
-            player.sendMessage('§4Invalid Team ID.');
-            return showPanel(player, panelId, context);
-        }
-
-        // Confirm Application
-        const { getTeam, applyToTeam } = await import('../teamManager.js');
-        const team = getTeam(teamId);
-        if (!team) {
-            player.sendMessage('§4Team not found.');
-            return showPanel(player, panelId, context);
-        }
-
-        showConfirmationDialog(player, {
-            title: `Apply to ${team.name}?`,
-            body: `Do you want to send a join request to ${team.name}?`,
-            confirmButtonText: '§2Apply',
-            cancelButtonText: 'Cancel',
-            onConfirm: () => {
-                const result = applyToTeam(player, teamId);
-                player.sendMessage(result.message || '§4Unknown error.');
-                showPanel(player, 'teamJoinPanel', context);
-            },
-            onCancel: () => showPanel(player, 'teamJoinPanel', context)
-        });
-        return;
-    }
-
-    if (panelId === 'teamInvitesPanel') {
-        const invites = (pData as PlayerData).pendingInvites || [];
-
-        if (selection === 0) {
-            return showPanel(player, 'teamJoinPanel', context);
-        } // Back
-
-        if (invites.length === 0) {
-            return;
-        } // Body text only
-
-        const denyAllIndex = invites.length + 1;
-
-        if (selection === denyAllIndex) {
-            const { updatePlayerData } = await import('../playerDataManager.js');
-            updatePlayerData(player.id, (d: PlayerData) => (d.pendingInvites = []));
-            player.sendMessage('§2Cleared all pending invites.');
-            return showPanel(player, panelId, context);
-        }
-
-        if (typeof selection !== 'number') return;
-        const inviteIndex = selection - 1;
-        if (inviteIndex >= 0 && inviteIndex < invites.length) {
-            const invite = invites[inviteIndex];
-            // Options: Accept, Deny
-            const { acceptInvite, denyInvite } = await import('../teamManager.js');
-
-            const form = new ActionFormData()
-                .title(`Invite: ${invite.teamName}`)
-                .button('§2Accept Invite', 'textures/ui/check')
-                .button('§4Deny Invite', 'textures/ui/cancel');
-
-            const res = await utils.uiWait(player, form);
-            if (!res.canceled) {
-                const resSelection = (res as ActionFormResponse).selection;
-                if (resSelection === 0) {
-                    // Accept
-                    const result = acceptInvite(player, invite.teamId);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                    if (result.success) {
-                        return showPanel(player, 'teamMainPanel', context);
-                    }
-                } else {
-                    // Deny
-                    const result = denyInvite(player.id, invite.teamId);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                }
-            }
-            return showPanel(player, panelId, context);
-        }
-        return;
-    }
-
-    if (panelId === 'teamBrowserPanel') {
-        const { getAllTeams } = await import('../teamManager.js');
-        const page = context.page || 1;
-        const teams = getAllTeams().sort((a, b) => b.members.length - a.members.length);
-        const paginatedTeams = getPaginatedItems(teams, page);
-
-        if (selection === 0) {
-            return showPanel(player, 'teamJoinPanel', context);
-        }
-
-        if (typeof selection !== 'number') return;
-        const selectionIndex = selection - 1;
-
-        if (selectionIndex < paginatedTeams.length) {
-            // Apply to selected team
-            const team = paginatedTeams[selectionIndex];
-            const { applyToTeam } = await import('../teamManager.js');
-
-            showConfirmationDialog(player, {
-                title: `Apply to ${team.name}?`,
-                body: `Do you want to send a join request to ${team.name}?`,
-                confirmButtonText: '§2Apply',
-                cancelButtonText: 'Cancel',
-                onConfirm: () => {
-                    const result = applyToTeam(player, team.id);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                    showPanel(player, panelId, context);
-                },
-                onCancel: () => showPanel(player, panelId, context)
-            });
-            return;
-        }
-
-        // Pagination logic
-        let buttonIndex = selectionIndex - paginatedTeams.length;
-        const totalPages = Math.ceil(teams.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page - 1 });
-        }
-        if (hasPrev) {
-            buttonIndex--;
-        }
-        if (hasNext && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-
-        return;
-    }
-
-    if (panelId === 'teamManagePanel') {
-        const { getTeamByPlayer, deleteTeam } = await import('../teamManager.js');
-        const team = getTeamByPlayer(player.id);
-        if (!team) {
-            return;
-        }
-
-        if (selection === 0) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-
-        if (selection === 1) {
-            // Invite Player
-            return showPanel(player, 'playerListPanel', { ...context, fromPanel: 'teamManagePanel' });
-        }
-
-        if (selection === 2) {
-            return showPanel(player, 'teamRequestsPanel', context);
-        }
-
-        if (selection === 3) {
-            // Manage Members
-            return showPanel(player, 'teamMembersPanel', { ...context, mode: 'manage' });
-        }
-
-        if (selection === 4) {
-            // Team Home
-            return showPanel(player, 'teamHomePanel', context);
-        }
-
-        if (selection === 5) {
-            // Delete Team (Owner Only)
-            showConfirmationDialog(player, {
-                title: 'Delete Team?',
-                body: '§4WARNING: This will disband the team and cannot be undone.',
-                confirmButtonText: '§4DELETE',
-                cancelButtonText: 'Cancel',
-                onConfirm: () => {
-                    deleteTeam(team.id);
-                    player.sendMessage('§2Team deleted.');
-                    showPanel(player, 'mainPanel', {});
-                },
-                onCancel: () => showPanel(player, panelId, context)
-            });
-            return;
-        }
-        return;
-    }
-
-    if (panelId === 'teamHomePanel') {
-        if (selection === 0) {
-            return showPanel(player, 'teamManagePanel', context);
-        }
-
-        const { getTeamByPlayer, setTeamHome, deleteTeamHome } = await import('../teamManager.js');
-        const team = getTeamByPlayer(player.id);
-        if (!team) {
-            return;
-        }
-
-        const isOwner = team.ownerId === player.id;
-
-        const isAdmin = team.admins.includes(player.id);
-        const canManage = isOwner || isAdmin;
-
-        let btnIndex = 1;
-
-        if (team.home) {
-            if (selection === btnIndex) {
-                // Teleport
-                const { x, y, z, dimensionId } = team.home;
-                const targetDimension = dimensionId
-                    ? (await import('@minecraft/server')).world.getDimension(dimensionId)
-                    : player.dimension;
-                if (targetDimension) {
-                    player.teleport({ x, y, z }, { dimension: targetDimension });
-                    player.sendMessage('§2Teleported to team home.');
-                } else {
-                    player.sendMessage('§4Invalid dimension for team home.');
-                }
-                return;
-            }
-            btnIndex++;
-        }
-
-        if (canManage) {
-            if (selection === btnIndex) {
-                // Update Location
-                setTeamHome(team.id, player.location, player.dimension.id);
-                player.sendMessage('§2Team home updated to your current location.');
-                return showPanel(player, panelId, context);
-            }
-            btnIndex++;
-
-            if (team.home) {
-                if (selection === btnIndex) {
-                    // Delete Home
-                    deleteTeamHome(team.id);
-                    player.sendMessage('§2Team home deleted.');
-                    return showPanel(player, panelId, context);
-                }
-            }
-        }
-        return;
-    }
-
-    if (panelId === 'teamRequestsPanel') {
-        const { getTeamByPlayer, acceptApplication, denyApplication } = await import('../teamManager.js');
-        const team = getTeamByPlayer(player.id);
-        if (!team) {
-            return;
-        }
-
-        if (selection === 0) {
-            return showPanel(player, 'teamManagePanel', context);
-        }
-
-        const appIndex = selection && selection > 0 ? selection - 1 : -1;
-        if (appIndex >= 0 && appIndex < team.applications.length) {
-            const app = team.applications[appIndex];
-
-            const form = new ActionFormData()
-                .title(`App: ${app.playerName}`)
-                .button('§2Accept', 'textures/ui/check')
-                .button('§4Deny', 'textures/ui/cancel');
-
-            const res = await utils.uiWait(player, form);
-            if (!res.canceled) {
-                const resSelection = (res as ActionFormResponse).selection;
-                if (resSelection === 0) {
-                    // Accept
-                    const result = acceptApplication(team.id, app.playerId);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                } else {
-                    // Deny
-                    const result = denyApplication(team.id, app.playerId);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                }
-            }
-            return showPanel(player, panelId, context);
-        }
-        return;
-    }
-
-    if (panelId === 'teamSettingsPanel') {
-        if (canceled) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-
-        const { getTeamByPlayer, setTeamOpenStatus } = await import('../teamManager.js');
-        const { updatePlayerData } = await import('../playerDataManager.js');
-
-        const team = getTeamByPlayer(player.id);
-        if (!team) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-
-        const isOwner = team.ownerId === player.id;
-
-        const isAdmin = team.admins.includes(player.id);
-        const canManage = isOwner || isAdmin;
-
-        const autoTp = (response as ModalFormResponse).formValues?.[0];
-
-        updatePlayerData(player.id, (d: PlayerData) => {
-            if (!d.teamSettings) {
-                d.teamSettings = { autoTpAccept: false };
-            }
-            d.teamSettings.autoTpAccept = autoTp as boolean;
-        });
-
-        if (canManage && formValues && formValues.length > 1) {
-            const allowRequests = formValues[1];
-            setTeamOpenStatus(team.id, allowRequests as boolean);
-        }
-
-        player.sendMessage('§2Settings updated.');
-        return showPanel(player, 'teamMainPanel', context);
-    }
-
-    if (panelId === 'teamMembersPanel') {
-        if (selection === 0) {
-            return showPanel(player, 'teamMainPanel', context);
-        }
-
-        const { getTeamByPlayer, kickMember, promoteMember, demoteMember, transferOwnership } =
-            await import('../teamManager.js');
-        const team = getTeamByPlayer(player.id);
-        if (!team) {
-            return;
-        }
-
-        if (typeof selection !== 'number') return;
-        const memberIndex = selection - 1;
-        if (memberIndex >= 0 && memberIndex < team.members.length) {
-            const memberId = team.members[memberIndex];
-
-            // Self-interaction check?
-            if (memberId === player.id) {
-                return showPanel(player, panelId, context);
-            }
-
-            const isOwner = team.ownerId === player.id;
-
-            const isAdmin = team.admins.includes(player.id);
-
-            // If viewing, maybe show profile? For now, if managing mode or high rank, show actions.
-            if (isOwner || isAdmin) {
-                const targetIsOwner = team.ownerId === memberId;
-
-                const targetIsAdmin = team.admins.includes(memberId);
-
-                // Access control
-                if (isAdmin && (targetIsOwner || targetIsAdmin)) {
-                    player.sendMessage('§4You cannot modify this member.');
-                    return showPanel(player, panelId, context);
-                }
-
-                const form = new ActionFormData().title('Manage Member');
-                form.button('Kick Member', 'textures/ui/cancel');
-                if (isOwner) {
-                    if (targetIsAdmin) {
-                        form.button('Demote to Member', 'textures/ui/arrow_down');
-                    } else {
-                        form.button('Promote to Admin', 'textures/ui/arrow_up');
-                    }
-
-                    form.button('Transfer Ownership', 'textures/ui/op');
-                }
-
-                const res = await utils.uiWait(player, form);
-                if (res.canceled) {
-                    return showPanel(player, panelId, context);
-                }
-
-                const resSelection = (res as ActionFormResponse).selection;
-
-                if (resSelection === 0) {
-                    // Kick
-                    const result = kickMember(team.id, memberId);
-                    player.sendMessage(result.message || '§4Unknown error.');
-                } else if (resSelection === 1) {
-                    // Promote/Demote
-                    if (isOwner) {
-                        if (targetIsAdmin) {
-                            const result = demoteMember(team.id, memberId);
-                            player.sendMessage(result.message || '§4Unknown error.');
-                        } else {
-                            const result = promoteMember(team.id, memberId);
-                            player.sendMessage(result.message || '§4Unknown error.');
-                        }
-                    }
-                } else if (resSelection === 2) {
-                    // Transfer
-                    showConfirmationDialog(player, {
-                        title: 'Transfer Ownership?',
-                        body: 'You will become a regular member.',
-                        confirmButtonText: '§4Confirm',
-                        cancelButtonText: 'Cancel',
-                        onConfirm: () => {
-                            const result = transferOwnership(team.id, memberId);
-                            player.sendMessage(result.message || '§4Unknown error.');
-                            showPanel(player, panelId, context);
-                        },
-                        onCancel: () => showPanel(player, panelId, context)
-                    });
-                    return;
-                }
-                return showPanel(player, panelId, context);
-            }
-        }
-        return showPanel(player, panelId, context);
-    }
-
-    if (panelId === 'xrayOresPanel') {
-        if (selection === 0) {
-            // Back
-            return showPanel(player, 'config_xray', context);
-        }
-        if (selection === 1) {
-            // Add New Ore
-            return showPanel(player, 'addXrayOrePanel', context);
-        }
-        const xrayConfig = getXrayConfig() as XrayConfig;
-        const ores = Object.values(xrayConfig.monitoredOreTypes || {}).sort(
-            (a: MonitoredOreType, b: MonitoredOreType) => a.oreName.localeCompare(b.oreName)
-        );
-        if (typeof selection === 'number') {
-            const selectedOreIndex = selection - 2;
-            if (selectedOreIndex >= 0 && selectedOreIndex < ores.length) {
-                return showPanel(player, 'editXrayOrePanel', { ...context, oreIndex: selectedOreIndex });
-            }
-        }
-        return;
-    }
-
-    if (panelId === 'addXrayOrePanel') {
-        if (canceled) {
-            return showPanel(player, 'xrayOresPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [blockId, dimensionId, minYStr, maxYStr, oreName] = values as string[];
-        const minY = parseInt(minYStr, 10);
-        const maxY = parseInt(maxYStr, 10);
-        if (blockId && dimensionId && !isNaN(minY) && !isNaN(maxY) && oreName) {
-            const xrayConfig = getXrayConfig() as XrayConfig;
-            // Define an extended type for mutable config locally since XrayConfig.monitoredOreTypes is likely partial or specific keys in default
-            const mutableConfig = xrayConfig as { monitoredOreTypes: Record<string, MonitoredOreType> };
-            if (!mutableConfig.monitoredOreTypes) {
-                mutableConfig.monitoredOreTypes = {};
-            }
-            const key = oreName.toLowerCase().replace(/\s+/g, '_');
-            mutableConfig.monitoredOreTypes[key] = {
-                enabled: true,
-                oreName,
-                blocks: [{ blockId, dimensionId, minY, maxY }]
-            };
-            saveXrayConfig(xrayConfig);
-            player.sendMessage('§2Successfully added new monitored ore.');
-        } else {
-            player.sendMessage('§4Invalid data. Please check all fields.');
-        }
-        return showPanel(player, 'xrayOresPanel', context);
-    }
-
-    if (panelId === 'editXrayOrePanel') {
-        if (canceled) {
-            return showPanel(player, 'xrayOresPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const { oreIndex } = context;
-        const [blockId, dimensionId, minYStr, maxYStr, oreName] = values as string[];
-        const minY = parseInt(minYStr, 10);
-        const maxY = parseInt(maxYStr, 10);
-        if (blockId && dimensionId && !isNaN(minY) && !isNaN(maxY) && oreName) {
-            const xrayConfig = getXrayConfig() as XrayConfig;
-            const mutableConfig = xrayConfig as { monitoredOreTypes: Record<string, MonitoredOreType> };
-
-            const oreTypes = mutableConfig.monitoredOreTypes;
-            const oreKeys = Object.keys(oreTypes || {}).sort((a, b) => {
-                const nameA = oreTypes[a].oreName;
-                const nameB = oreTypes[b].oreName;
-                return nameA.localeCompare(nameB);
-            });
-            const key = oreKeys[oreIndex];
-
-            if (key && oreTypes && oreTypes[key]) {
-                oreTypes[key] = {
-                    enabled: true,
-                    oreName,
-                    blocks: [{ blockId, dimensionId, minY, maxY }]
-                };
-                saveXrayConfig(xrayConfig);
-                player.sendMessage('§2Successfully updated monitored ore.');
-            }
-        } else {
-            player.sendMessage('§4Invalid data. Please check all fields.');
-        }
-        return showPanel(player, 'xrayOresPanel', context);
-    }
-
-    if (panelId === 'floatingTextEditPanel') {
-        if (canceled) {
-            return showPanel(player, 'floatingTextActionPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const { id } = context;
-        const [textContent, x, y, z, dimensionIndex, useExpiration, expirationMinutes] = values as [
-            string,
-            string,
-            string,
-            string,
-            number,
-            boolean,
-            string
-        ];
-
-        const dimensionIds = ['minecraft:overworld', 'minecraft:nether', 'minecraft:the_end'];
-        const selectedDimension = dimensionIds[dimensionIndex] ?? 'minecraft:overworld';
-
-        const updatedConfig = {
-            text: textContent,
-            location: { x: parseFloat(x), y: parseFloat(y), z: parseFloat(z) },
-            dimension: selectedDimension,
-            expiresAt:
-                useExpiration && Number(expirationMinutes) > 0 ? Date.now() + Number(expirationMinutes) * 60000 : null
-        };
-        floatingTextManager.updateText(id, updatedConfig);
-        player.sendMessage(`§2Successfully updated floating text: ${id}`);
-        return showPanel(player, 'floatingTextActionPanel', context);
-    }
-
-    if (panelId === 'floatingTextCreatePanel') {
-        if (canceled) {
-            return showPanel(player, 'floatingTextListPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [id, text] = values as string[];
-        if (!id) {
-            player.sendMessage('§4ID cannot be empty.');
-            return showPanel(player, 'floatingTextCreatePanel', context);
-        }
-        if (id.includes(' ')) {
-            player.sendMessage('§4ID cannot contain spaces. Please use a single word.');
-            return showPanel(player, 'floatingTextCreatePanel', context);
-        }
-        if (floatingTextManager.createText(player, id, text)) {
-            // Success message is sent by createText
-        }
-        return showPanel(player, 'floatingTextListPanel', context);
-    }
-
-    if (panelId === 'rulesManagementPanel') {
-        const page = context.page || 1;
-        const rules = rulesManager.getRules();
-
-        // Back button
-        if (selection === 0) {
-            return showPanel(player, 'infoPanel', context);
-        }
-
-        let offset = 1;
-        if (pData.permissionLevel <= 1) {
-            // "Add Rule" button
-            if (selection === 1) {
-                return showPanel(player, 'addRulePanel', context);
-            }
-            offset++;
-        }
-
-        const paginatedRules = getPaginatedItems(rules, page);
-        if (typeof selection !== 'number') return;
-        const selectionIndex = selection - offset;
-
-        // Handle rule selection
-        if (selectionIndex < paginatedRules.length) {
-            const ruleIndex = (page - 1) * itemsPerPage + selectionIndex;
-            return showPanel(player, 'ruleActionPanel', { ...context, ruleIndex });
-        }
-
-        // Handle pagination
-        let buttonIndex = selectionIndex - paginatedRules.length;
-        const totalPages = Math.ceil(rules.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (rules.length > itemsPerPage) {
-            if (hasPrev && buttonIndex === 0) {
-                return showPanel(player, panelId, { ...context, page: page - 1 });
-            }
-            if (hasPrev) {
-                buttonIndex--;
-            }
-
-            if (hasNext && buttonIndex === 0) {
-                return showPanel(player, panelId, { ...context, page: page + 1 });
-            }
-            if (hasNext) {
-                buttonIndex--;
-            }
-        }
-        return;
-    }
-
-    if (panelId === 'addRulePanel') {
-        if (canceled) {
-            return showPanel(player, 'rulesManagementPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [newRuleText] = values as string[];
-        if (newRuleText) {
-            rulesManager.addRule(newRuleText);
-            player.sendMessage('§2Rule added successfully.');
-        }
-        return showPanel(player, 'rulesManagementPanel', context);
-    }
-
-    if (panelId === 'helpfulLinksManagementPanel') {
-        const page = context.page || 1;
-        const links = helpfulLinksManager.getHelpfulLinks();
-
-        // Back button
-        if (selection === 0) {
-            return showPanel(player, 'infoPanel', context);
-        }
-
-        let offset = 1;
-        if (pData.permissionLevel <= 1) {
-            // Add Link button
-            if (selection === 1) {
-                return showPanel(player, 'addHelpfulLinkPanel', context);
-            }
-            offset++;
-        }
-
-        const paginatedLinks = getPaginatedItems(links, page);
-        if (typeof selection !== 'number') return;
-        const selectionIndex = selection - offset;
-
-        // Handle link selection
-        if (selectionIndex < paginatedLinks.length) {
-            const linkIndex = (page - 1) * itemsPerPage + selectionIndex;
-            return showPanel(player, 'helpfulLinkActionPanel', { ...context, linkIndex });
-        }
-
-        // Handle pagination
-        let buttonIndex = selectionIndex - paginatedLinks.length;
-        const totalPages = Math.ceil(links.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (links.length > itemsPerPage) {
-            if (hasPrev && buttonIndex === 0) {
-                return showPanel(player, panelId, { ...context, page: page - 1 });
-            }
-            if (hasPrev) {
-                buttonIndex--;
-            }
-
-            if (hasNext && buttonIndex === 0) {
-                return showPanel(player, panelId, { ...context, page: page + 1 });
-            }
-            if (hasNext) {
-                buttonIndex--;
-            }
-        }
-        return;
-    }
-
-    if (panelId === 'addHelpfulLinkPanel') {
-        if (canceled) {
-            return showPanel(player, 'helpfulLinksManagementPanel', context);
-        }
-        const values = (response as ModalFormResponse).formValues;
-        if (!values) return;
-
-        const [title, url] = values as string[];
-        if (title && url) {
-            helpfulLinksManager.addHelpfulLink(title, url);
-            player.sendMessage('§2Link added successfully.');
-        }
-        return showPanel(player, 'helpfulLinksManagementPanel', context);
-    }
-
-    if (panelId === 'helpfulLinkActionPanel') {
-        const { linkIndex } = context;
-        if (typeof selection !== 'number') return;
-        switch (selection) {
-            case 0: {
-                // Edit
-                const links = helpfulLinksManager.getHelpfulLinks();
-                const currentLink = links[linkIndex];
-                const editForm = new ModalFormData()
-                    .title('Edit Link')
-                    .textField('Link Title', 'Enter the new title', { defaultValue: currentLink.title })
-                    .textField('Link URL', 'Enter the new URL', { defaultValue: currentLink.url });
-                const editResponse = await utils.uiWait(player, editForm);
-                if (editResponse.canceled) {
-                    return showPanel(player, 'helpfulLinkActionPanel', context);
-                }
-                const values = (editResponse as ModalFormResponse).formValues;
-                if (!values) return;
-
-                const [newTitle, newUrl] = values as string[];
-                if (newTitle && newUrl) {
-                    helpfulLinksManager.editHelpfulLink(linkIndex, newTitle, newUrl);
-                    player.sendMessage('§2Link updated successfully.');
-                }
-                return showPanel(player, 'helpfulLinksManagementPanel', context);
-            }
-            case 1: // Move Up
-                helpfulLinksManager.moveHelpfulLink(linkIndex, 'up');
-                return showPanel(player, 'helpfulLinksManagementPanel', context);
-            case 2: // Move Down
-                helpfulLinksManager.moveHelpfulLink(linkIndex, 'down');
-                return showPanel(player, 'helpfulLinksManagementPanel', context);
-            case 3: {
-                // Delete Link
-                showConfirmationDialog(player, {
-                    title: '§4Confirm Deletion',
-                    body: 'Are you sure you want to delete this link?',
-                    confirmButtonText: '§4Yes, Delete',
-                    cancelButtonText: '§2No, Cancel',
-                    onConfirm: () => {
-                        helpfulLinksManager.deleteHelpfulLink(linkIndex);
-                        player.sendMessage('§2Link deleted successfully.');
-                        showPanel(player, 'helpfulLinksManagementPanel', context);
-                    },
-                    onCancel: () => {
-                        showPanel(player, 'helpfulLinksManagementPanel', context);
-                    }
-                });
-                return;
-            }
-            case 4: // Back
-                return showPanel(player, 'helpfulLinksManagementPanel', context);
-        }
-        return;
-    }
-
-    if (panelId === 'ruleActionPanel') {
-        const { ruleIndex } = context;
-
-        if (typeof selection !== 'number') return;
-        switch (selection) {
-            case 0: {
-                // Edit Text
-                const rules = rulesManager.getRules();
-                const currentText = rules[ruleIndex];
-                const editForm = new ModalFormData()
-                    .title('Edit Rule Text')
-                    .textField('Rule text', 'Enter the new text', { defaultValue: currentText });
-
-                const editResponse = await utils.uiWait(player, editForm);
-                if (editResponse.canceled) {
-                    return showPanel(player, 'ruleActionPanel', context);
-                }
-
-                const values = (editResponse as ModalFormResponse).formValues;
-                if (!values) return;
-
-                const [newText] = values as string[];
-                if (newText) {
-                    rulesManager.editRule(ruleIndex, newText);
-                    player.sendMessage('§2Rule updated successfully.');
-                }
-                return showPanel(player, 'rulesManagementPanel', context);
-            }
-            case 1: // Move Up
-                rulesManager.moveRule(ruleIndex, 'up');
-                return showPanel(player, 'rulesManagementPanel', context);
-            case 2: // Move Down
-                rulesManager.moveRule(ruleIndex, 'down');
-                return showPanel(player, 'rulesManagementPanel', context);
-            case 3: {
-                // Delete Rule
-                showConfirmationDialog(player, {
-                    title: '§4Confirm Deletion',
-                    body: 'Are you sure you want to delete this rule?',
-                    confirmButtonText: '§4Yes, Delete',
-                    cancelButtonText: '§2No, Cancel',
-                    onConfirm: () => {
-                        rulesManager.deleteRule(ruleIndex);
-                        player.sendMessage('§2Rule deleted successfully.');
-                        showPanel(player, 'rulesManagementPanel', context);
-                    },
-                    onCancel: () => {
-                        showPanel(player, 'rulesManagementPanel', context);
-                    }
-                });
-                return;
-            }
-            case 4: // Back
-                return showPanel(player, 'rulesManagementPanel', context);
-        }
-        return;
-    }
-
-    // --- Shop Panel Handlers ---
-    if (panelId === 'shopMainPanel') {
-        if (selection === 0) {
-            return showPanel(player, 'mainPanel');
-        }
-        if (typeof selection !== 'number') return;
-
-        const shopConfig = getShopConfig();
-
-        const validCategories = Object.keys(shopConfig.categories)
-            .filter((categoryName: string) => {
-                const category = shopConfig.categories[categoryName];
-                const hasItems = Object.keys(category.items).length > 0;
-                const hasSubCategories = Object.keys(category.subCategories).length > 0;
-                return hasItems || hasSubCategories;
-            })
-            .sort();
-        const selectedCategoryName = validCategories[selection - 1];
-        if (selectedCategoryName) {
-            return showPanel(player, `shopCategoryPanel_${selectedCategoryName}`, {
-                ...context,
-                categoryName: selectedCategoryName
-            });
-        }
-        return;
-    }
-
-    if (panelId === 'configCategoryPanel') {
-        const page = context.page || 1;
-        if (typeof selection !== 'number') return;
-        if (selection === 0) return showPanel(player, 'mainPanel', context);
-
-        const sortedSystems = getVisibleConfigSystems(pData);
-        const paginatedSystems = getPaginatedItems(sortedSystems, page);
-
-        let buttonIndex = selection - 1;
-
-        if (buttonIndex >= 0 && buttonIndex < paginatedSystems.length) {
-            const system = paginatedSystems[buttonIndex];
-            return showPanel(player, system.id, context);
-        }
-        buttonIndex -= paginatedSystems.length;
-
-        const totalPages = Math.ceil(sortedSystems.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-
-        if (hasPrev) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page - 1 });
-            buttonIndex--;
-        }
-        if (hasNext) {
-            if (buttonIndex === 0) return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-        return;
-    }
-
-    if (panelId.startsWith('config_')) {
-        if (canceled) {
-            return showPanel(player, 'configCategoryPanel', context);
-        }
-
-        const categoryId = panelId.replace('config_', '');
-        const category = configPanelSchema.find((c) => c.id === categoryId);
-        if (category) {
-            const values = (response as ModalFormResponse).formValues;
-            if (values) {
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                const updates: Record<string, any> = {};
-                category.settings.forEach((setting, index) => {
-                    let value = values[index];
-                    if (setting.type === 'dropdown') {
-                        const options = setting.options || [];
-                        const selectedIndex = value as number;
-                        if (setting.key === 'logLevel') {
-                            value = selectedIndex;
-                        } else {
-                            value = options[selectedIndex];
-                        }
-                    } else if (setting.type === 'textField') {
-                        const strVal = value as string;
-                        const current = getValueFromPath(getConfig(), setting.key);
-                        if (typeof current === 'number' && !isNaN(Number(strVal)) && strVal.trim() !== '') {
-                            value = Number(strVal);
-                        }
-                    }
-                    updates[setting.key] = value;
-                });
-
-                const configSource = category.configSource || 'main';
-                const handler = configHandlers[configSource];
-                if (handler) {
-                    if (configSource === 'main') {
-                        handler.save(updates);
-                    } else {
-                        const currentConfig = handler.get();
-                        for (const key in updates) {
-                            setValueByPath(currentConfig, key, updates[key]);
-                        }
-                        handler.save(currentConfig);
-                    }
-                    player.sendMessage('§2Configuration saved.');
-
-                    if (categoryId === 'data') {
-                        import('../dataManager.js').then(({ restartAutoSave }) => restartAutoSave());
-                    }
-                }
-            }
-        }
-        return showPanel(player, 'configCategoryPanel', context);
-    }
-
-    if (panelId === 'configResetPanel') {
-        const page = context.page || 1;
-        const resettableSystems = [
-            ...configPanelSchema
-                .filter((c) => !c.id.startsWith('general_')) // General settings are not individually resettable via this panel
-                .map((c) => ({ id: c.id, title: c.title, icon: c.icon })),
-            { id: 'kits', title: '§l§5Kit System§r', icon: 'textures/ui/inventory_icon' },
-            { id: 'shop', title: '§l§2Shop System§r', icon: 'textures/items/emerald' },
-            { id: 'ranks', title: '§l§4Rank System§r', icon: 'textures/ui/permissions_member_star.png' }
-        ];
-        resettableSystems.sort((a, b) => a.title.replace(/§./g, '').localeCompare(b.title.replace(/§./g, '')));
-
-        const sortedSystems = resettableSystems;
-
-        if (selection === 0) {
-            // Back button
-            return showPanel(player, 'configCategoryPanel', { ...context, page: 1 });
-        }
-
-        const paginatedSystems = getPaginatedItems(sortedSystems, page);
-        const selectionIndex = selection && selection > 0 ? selection - 1 : -1;
-
-        if (selectionIndex >= 0 && selectionIndex < paginatedSystems.length) {
-            const selectedSystem = paginatedSystems[selectionIndex];
-            showConfirmationDialog(player, {
-                title: `Confirm Reset: ${selectedSystem.title}`,
-                body: `This action cannot be undone. Are you sure you want to reset the ${selectedSystem.title} configuration to its default values?`,
-                confirmButtonText: '§4Yes, Reset',
-                cancelButtonText: '§2No, Cancel',
-                onConfirm: async () => {
-                    const finalConfirmForm = new ModalFormData()
-                        .title('Final Confirmation')
-                        .textField(`Type "confirm" to reset ${selectedSystem.title}.`, 'Case-insensitive', {
-                            defaultValue: ''
-                        });
-
-                    const finalConfirmResponse = await utils.uiWait(player, finalConfirmForm);
-
-                    if (finalConfirmResponse.canceled) {
-                        player.sendMessage('§4Final confirmation failed. Reset canceled.');
-                        return showPanel(player, 'configResetPanel', { ...context, page });
-                    }
-
-                    const confirmModal = finalConfirmResponse as ModalFormResponse;
-                    const confirmationValue =
-                        confirmModal.formValues && confirmModal.formValues[0] ? String(confirmModal.formValues[0]) : '';
-
-                    if (confirmationValue.trim().toLowerCase() !== 'confirm') {
-                        player.sendMessage('§4Final confirmation failed. Reset canceled.');
-                        return showPanel(player, 'configResetPanel', { ...context, page });
-                    }
-
-                    const result = await resetConfigSection(selectedSystem.id, player);
-                    if (result.success) {
-                        player.sendMessage(`§2${result.message}`);
-                    } else {
-                        player.sendMessage(
-                            '§4Failed to reset the configuration. Please check the console for details.'
-                        );
-                        errorLog(
-                            `[UIManager] Failed to reset config section '${selectedSystem.id}': ${result.message}`
-                        );
-                    }
-                    return showPanel(player, 'configResetPanel', { ...context, page: 1 });
-                },
-                onCancel: () => {
-                    player.sendMessage('§2Reset canceled.');
-                    return showPanel(player, 'configResetPanel', { ...context, page });
-                }
-            });
-            return;
-        }
-
-        const buttonIndex = selectionIndex >= 0 ? selectionIndex - paginatedSystems.length : -1;
-
-        const totalPages = Math.ceil(resettableSystems.length / itemsPerPage);
-
-        if (page >= totalPages && buttonIndex === 0) {
-            showConfirmationDialog(player, {
-                title: 'Confirm Reset All',
-                body: 'This action cannot be undone. Are you sure you want to reset ALL system configurations to their default values?',
-                confirmButtonText: '§4Yes, Reset All',
-                cancelButtonText: '§2No, Cancel',
-                onConfirm: async () => {
-                    const finalConfirmForm = new ModalFormData()
-                        .title('Final Confirmation')
-                        .textField('Type "confirm" to reset ALL systems.', 'Case-insensitive', {
-                            defaultValue: ''
-                        });
-
-                    const finalConfirmResponse = await utils.uiWait(player, finalConfirmForm);
-
-                    if (finalConfirmResponse.canceled) {
-                        player.sendMessage('§4Final confirmation failed. Reset canceled.');
-                        return showPanel(player, 'configResetPanel', { ...context, page });
-                    }
-
-                    const confirmModal = finalConfirmResponse as ModalFormResponse;
-                    const confirmationValue =
-                        confirmModal.formValues && confirmModal.formValues[0] ? String(confirmModal.formValues[0]) : '';
-
-                    if (confirmationValue.trim().toLowerCase() !== 'confirm') {
-                        player.sendMessage('§4Final confirmation failed. Reset canceled.');
-                        return showPanel(player, 'configResetPanel', { ...context, page });
-                    }
-
-                    const result = await resetConfigSection('all', player);
-                    if (result.success) {
-                        player.sendMessage(`§2${result.message}`);
-                    } else {
-                        player.sendMessage(
-                            '§4Failed to reset all configurations. Please check the console for details.'
-                        );
-                        errorLog(`[UIManager] Failed to reset all config sections: ${result.message}`);
-                    }
-                    return showPanel(player, 'configResetPanel', { ...context, page: 1 });
-                },
-                onCancel: () => {
-                    player.sendMessage('§2Reset canceled.');
-                    return showPanel(player, 'configResetPanel', { ...context, page });
-                }
-            });
-            return;
-        }
-
-        // Handle pagination
-        const hasPrev = page > 1;
-
-        if (hasPrev && buttonIndex === 0) {
-            return showPanel(player, panelId, { ...context, page: page - 1 });
-        }
-        if (buttonIndex >= 0) {
-            // Should be next page
-            return showPanel(player, panelId, { ...context, page: page + 1 });
-        }
-
-        return;
-    }
-
-    if (panelId.startsWith('shopCategoryPanel_') || panelId.startsWith('shopItemListPanel_')) {
-        const isItemList = panelId.startsWith('shopItemListPanel_');
-        const prefix = isItemList ? 'shopItemListPanel_' : 'shopCategoryPanel_';
-        const rawId = panelId.replace(prefix, '');
-        const parts = rawId.split('_');
-        const categoryName = parts[0];
-        const subCategoryName = isItemList ? parts.slice(1).join('_') : undefined;
-        const page = context.page || 1;
-        const view = context.view || 'shop';
-
-        if (selection === 0) {
-            // Back button
-            const parentPanel = isItemList ? `shopCategoryPanel_${categoryName}` : 'shopMainPanel';
-            return showPanel(player, parentPanel, { ...context, page: 1 });
-        }
-
-        // Reconstruct the list of entries that was shown to the player
-        const shopConfig = getShopConfig();
-        const category = shopConfig.categories[categoryName];
-        let allEntries: { type: string }[] = [];
-        if (isItemList && subCategoryName) {
-            const subCategory = category.subCategories[subCategoryName];
-            allEntries = Object.keys(subCategory.items).map((id) => ({ id, ...subCategory.items[id], type: 'item' }));
-        } else {
-            // shopCategoryPanel
-            const subCategories = Object.keys(category.subCategories)
-                .sort()
-                .map((name) => ({ name, ...category.subCategories[name], type: 'subCategory' }));
-            const items = Object.keys(category.items).map((id) => ({ id, ...category.items[id], type: 'item' }));
-            allEntries = [...subCategories, ...items];
-        }
-
-        const paginatedEntries = getPaginatedItems(allEntries, page);
-        const selectionIndex = selection && selection > 0 ? selection - 1 : -1;
-
-        // Handle pagination
-        if (selectionIndex >= paginatedEntries.length) {
-            let newPage = page;
-            const totalPages = Math.ceil(allEntries.length / itemsPerPage);
-            const hasPrev = page > 1;
-            const hasNext = page < totalPages;
-            const buttonIndex = selectionIndex - paginatedEntries.length;
-
-            if (hasPrev && buttonIndex === 0) {
-                newPage--;
-            } else if (hasNext) {
-                newPage++;
-            }
-            return showPanel(player, panelId, { ...context, page: newPage });
-        }
-
-        const selectedEntry =
-            selectionIndex >= 0
-                ? (paginatedEntries[selectionIndex] as {
-                      type: string;
-                      name: string;
-                      id: string;
-                      buyPrice: number;
-                      sellPrice: number;
-                  })
-                : undefined;
-
-        if (selectedEntry && selectedEntry.type === 'subCategory') {
-            return showPanel(player, `shopItemListPanel_${categoryName}_${selectedEntry.name}`, {
-                ...context,
-                categoryName,
-                subCategoryName: selectedEntry.name,
-                page: 1
-            });
-        }
-
-        // It's an item
-        if (selectedEntry) {
-            const itemId = selectedEntry.id;
-            const masterItem = allItems[itemId];
-            const shopItem = selectedEntry;
-
-            const canBuy = view !== 'sell' && shopItem.buyPrice > 0;
-            const canSell = view !== 'buy' && shopItem.sellPrice > 0;
-
-            if (!canBuy && !canSell) {
-                player.sendMessage('§4This item cannot be bought or sold currently.');
-                return showPanel(player, panelId, context);
-            }
-
-            const modal = new ModalFormData().title(masterItem.displayName ?? itemId);
-            let action;
-            let hasDropdown = false;
-
-            if (canBuy && canSell) {
-                modal.textField('Amount', 'Enter the amount', { defaultValue: '1' });
-                const options = [`Buy ($${shopItem.buyPrice})`, `Sell ($${shopItem.sellPrice})`];
-                modal.dropdown('Action', options, { defaultValueIndex: 0 });
-                hasDropdown = true;
-            } else if (canBuy) {
-                modal.textField(`Amount to Buy (Price: $${shopItem.buyPrice})`, 'Enter a numeric value', {
-                    defaultValue: '1'
-                });
-                action = 'buy';
-            } else {
-                // canSell
-                modal.textField(`Amount to Sell (Price: $${shopItem.sellPrice})`, 'Enter a numeric value', {
-                    defaultValue: '1'
-                });
-                action = 'sell';
-            }
-
-            const modalResponse = await utils.uiWait(player, modal);
-
-            if (modalResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-
-            let amount;
-            if (hasDropdown) {
-                const values = (modalResponse as ModalFormResponse).formValues as [string, number];
-                const amountStr = values[0];
-                const actionIndex = values[1];
-                amount = parseInt(amountStr, 10);
-                const options = [`Buy ($${shopItem.buyPrice})`, `Sell ($${shopItem.sellPrice})`];
-                const selectedActionString = options[actionIndex];
-                action = selectedActionString.startsWith('Buy') ? 'buy' : 'sell';
-            } else {
-                const values = (modalResponse as ModalFormResponse).formValues as [string];
-                const amountStr = values[0];
-                amount = parseInt(amountStr, 10);
-            }
-
-            if (isNaN(amount) || amount <= 0) {
-                player.sendMessage('§4Invalid amount.');
-                return showPanel(player, panelId, context);
-            }
-
-            let result;
-            if (action === 'buy') {
-                result = shopManager.buyItem(player, itemId, amount);
-            } else {
-                // action === 'sell'
-                result = shopManager.sellItem(player, itemId, amount);
-            }
-            player.sendMessage(result.message);
-
-            return showPanel(player, panelId, context); // Refresh the panel
-        }
-    }
-
-    // --- Admin Edit Shop Panel Handlers ---
-    if (panelId.startsWith('shopAddItemPanel_')) {
-        const { categoryName, page = 1 } = context;
-        if (selection === 0) {
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, context);
-        }
-
-        if (selection === 1) {
-            // Add Custom Item
-            const form = new ModalFormData()
-                .title('Add Custom Item')
-                .textField('Item ID (unique key)', 'e.g., custom_sword')
-                .textField('Display Name', 'e.g., Sword of Awesome')
-                .textField('Minecraft Item ID', 'e.g., minecraft:diamond_sword')
-                .textField('Icon Path', 'e.g., textures/items/diamond_sword')
-                .textField('Buy Price', '-1 to disable', { defaultValue: '-1' })
-                .textField('Sell Price', '-1 to disable', { defaultValue: '-1' })
-                .textField('Permission Level', 'e.g., 1024', { defaultValue: '1024' });
-            const addItemResponse = await utils.uiWait(player, form);
-            if (addItemResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-            const values = (addItemResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const customId = values[0] as string;
-            const displayName = values[1] as string;
-            const mcId = values[2] as string;
-            const iconStr = values[3] as string;
-            const buyPriceStr = values[4] as string;
-            const sellPriceStr = values[5] as string;
-            const permLevelStr = values[6] as string;
-
-            const icon = iconStr || '';
-            const buyPrice = parseInt(buyPriceStr, 10);
-            const sellPrice = parseInt(sellPriceStr, 10);
-            const permissionLevel = parseInt(permLevelStr, 10);
-
-            if (
-                customId &&
-                displayName &&
-                mcId &&
-                icon &&
-                !isNaN(buyPrice) &&
-                !isNaN(sellPrice) &&
-                !isNaN(permissionLevel)
-            ) {
-                shopAdminManager.addCustomItemToConfig(customId, {
-                    itemId: mcId,
-                    icon,
-                    buyPrice,
-                    sellPrice,
-                    displayName
-                });
-                shopAdminManager.setItem(categoryName, null, customId, {
-                    buyPrice,
-                    sellPrice,
-                    permissionLevel,
-                    icon,
-                    displayName,
-                    itemId: customId // ADDED
-                });
-                player.sendMessage(`§2Successfully added custom item '${displayName}'.`);
-            } else {
-                player.sendMessage('§4Invalid custom item data.');
-            }
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, { ...context, page: 1 });
-        }
-
-        const allPossibleItems = Object.keys(allItems);
-        const paginatedItems = getPaginatedItems(allPossibleItems, page);
-        const selectedItemId = selection && selection > 1 ? paginatedItems[selection - 2] : undefined;
-
-        if (selectedItemId) {
-            const masterItem = allItems[selectedItemId];
-            const form = new ModalFormData()
-                .title(`Add ${masterItem.displayName}`)
-                .textField('Icon Path', 'e.g., textures/items/diamond_sword', { defaultValue: masterItem.icon })
-                .textField('Buy Price', '-1 to disable', { defaultValue: `${masterItem.buyPrice}` })
-                .textField('Sell Price', '-1 to disable', { defaultValue: `${masterItem.sellPrice}` })
-                .textField('Permission Level', 'e.g., 1024', { defaultValue: '1024' });
-            const addItemResponse = await utils.uiWait(player, form);
-            if (addItemResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-            const values = (addItemResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const iconStr = values[0] as string;
-            const buyPriceStr = values[1] as string;
-            const sellPriceStr = values[2] as string;
-            const permLevelStr = values[3] as string;
-
-            const icon = iconStr || '';
-            const buyPrice = parseInt(buyPriceStr, 10);
-            const sellPrice = parseInt(sellPriceStr, 10);
-            const permissionLevel = parseInt(permLevelStr, 10);
-            if (!isNaN(buyPrice) && !isNaN(sellPrice) && !isNaN(permissionLevel)) {
-                const result = shopAdminManager.setItem(categoryName, null, selectedItemId, {
-                    buyPrice,
-                    sellPrice,
-                    permissionLevel,
-                    icon,
-                    itemId: selectedItemId, // ADDED
-                    displayName: masterItem.displayName || '' // ADDED
-                });
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, { ...context, page: 1 });
-        }
-        // Handle pagination
-        let newPage = page;
-        const totalPages = Math.ceil(allPossibleItems.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-        const buttonIndex = selection && selection > 1 ? selection - 2 - paginatedItems.length : -1;
-
-        if (hasPrev && buttonIndex === 0) {
-            newPage--;
-        } else if (hasNext) {
-            newPage++;
-        }
-        return showPanel(player, panelId, { ...context, page: newPage });
-    }
-
-    if (panelId === 'shopManagementPanel') {
-        const page = context.page || 1;
-        if (selection === 0) {
-            return showPanel(player, 'configCategoryPanel');
-        }
-
-        if (selection === 1) {
-            const mainConfig = getConfig();
-            const newStatus = !mainConfig.shop.enabled;
-            updateMultipleConfig({ 'shop.enabled': newStatus });
-            player.sendMessage(`§2Shop system has been ${newStatus ? 'enabled' : 'disabled'}.`);
-            return showPanel(player, 'shopManagementPanel', { ...context, page: 1 });
-        }
-
-        if (selection === 2) {
-            // Add Category
-            const form = new ModalFormData()
-                .title('Add Category')
-                .textField('Category Name', 'Enter category name', { defaultValue: '' })
-                .textField('Icon', 'Enter icon texture path', { defaultValue: '' });
-            const addCatResponse = await utils.uiWait(player, form);
-            if (addCatResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-            const values = (addCatResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const name = values[0] as string;
-            const iconStr = values[1] as string;
-
-            if (name) {
-                const result = shopAdminManager.addCategory(name, iconStr || '');
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, panelId, { ...context, page: 1 });
-        }
-
-        const shopConfig = getShopConfig();
-        const categories = Object.keys(shopConfig.categories).sort();
-        const paginatedCategories = getPaginatedItems(categories, page);
-        const selectedCategoryName = selection && selection > 2 ? paginatedCategories[selection - 3] : undefined;
-
-        if (selectedCategoryName) {
-            return showPanel(player, `shopAdminCategoryPanel_${selectedCategoryName}`, {
-                categoryName: selectedCategoryName
-            });
-        }
-        // Handle pagination
-        let newPage = page;
-        const totalPages = Math.ceil(categories.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-        const buttonIndex = selection && selection > 2 ? selection - 3 - paginatedCategories.length : -1;
-
-        if (hasPrev && buttonIndex === 0) {
-            newPage--;
-        } else if (hasNext) {
-            newPage++;
-        }
-        return showPanel(player, panelId, { ...context, page: newPage });
-    }
-
-    if (panelId.startsWith('shopAdminCategoryPanel_')) {
-        const { categoryName, page = 1 } = context;
-        if (selection === 0) {
-            return showPanel(player, 'shopManagementPanel');
-        }
-        if (selection === 1) {
-            // Add Item
-            return showPanel(player, `shopAddItemPanel_${categoryName}`, context);
-        }
-        if (selection === 2) {
-            // Add Subcategory
-            const form = new ModalFormData()
-                .title('Add Subcategory')
-                .textField('Subcategory Name', 'Enter subcategory name', { defaultValue: '' })
-                .textField('Icon', 'Enter icon texture path', { defaultValue: '' });
-            const addSubCatResponse = await utils.uiWait(player, form);
-            if (addSubCatResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-            const values = (addSubCatResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const name = values[0] as string;
-            const iconStr = values[1] as string;
-
-            if (name) {
-                const result = shopAdminManager.addSubCategory(categoryName, name, iconStr || '');
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, panelId, { ...context, page: 1 });
-        }
-        if (selection === 3) {
-            // Edit Category
-            return showPanel(player, `shopAdminCategoryActionPanel_${categoryName}`, context);
-        }
-
-        const shopConfig = getShopConfig();
-        const category = shopConfig.categories[categoryName];
-        const subCategories = Object.keys(category.subCategories)
-            .sort()
-            .map((name) => ({ name, ...category.subCategories[name], type: 'subCategory' }));
-        const items = Object.keys(category.items).map((id) => ({ id, ...category.items[id], type: 'item' }));
-        const allEntries = [...subCategories, ...items];
-        const paginatedEntries = getPaginatedItems(allEntries, page);
-        const selectedEntry =
-            selection && selection > 3
-                ? (paginatedEntries[selection - 4] as {
-                      type: string;
-                      id: string;
-                      displayName: string;
-                      icon: string;
-                      buyPrice: number;
-                      sellPrice: number;
-                      permissionLevel: number;
-                      name: string;
-                  })
-                : undefined;
-
-        if (selectedEntry) {
-            if (selectedEntry.type === 'item') {
-                const form = new ActionFormData()
-                    .title('Edit Item')
-                    .button('Edit', 'textures/ui/icon_setting')
-                    .button('Delete', 'textures/ui/trash');
-                const itemActionResponse = await utils.uiWait(player, form);
-                if (itemActionResponse.canceled) {
-                    return showPanel(player, panelId, context);
-                }
-                const itemActionSelection = (itemActionResponse as ActionFormResponse).selection;
-                if (itemActionSelection === 0) {
-                    // Edit
-                    const masterItem = allItems[selectedEntry.id] || {};
-                    const editForm = new ModalFormData()
-                        .title(`Edit Item: ${selectedEntry.id}`)
-                        .textField('Display Name', 'e.g., Magical Sword', {
-                            defaultValue: selectedEntry.displayName || masterItem.displayName
-                        })
-                        .textField('Minecraft Item ID', 'e.g., minecraft:diamond_sword', {
-                            defaultValue: masterItem.itemId
-                        })
-                        .textField('Icon Path', 'e.g., textures/items/diamond_sword', {
-                            defaultValue: selectedEntry.icon || masterItem.icon
-                        })
-                        .textField('Buy Price', '-1 to disable', { defaultValue: String(selectedEntry.buyPrice) })
-                        .textField('Sell Price', '-1 to disable', { defaultValue: String(selectedEntry.sellPrice) })
-                        .textField('Permission Level', 'e.g., 1024', {
-                            defaultValue: String(selectedEntry.permissionLevel)
-                        });
-
-                    const editResponse = await utils.uiWait(player, editForm);
-                    if (editResponse.canceled) {
-                        return showPanel(player, panelId, context);
-                    }
-
-                    const values = (editResponse as ModalFormResponse).formValues as string[];
-                    const [displayName, minecraftId, iconStr, buyPriceStr, sellPriceStr, permLevelStr] = values;
-                    const icon = iconStr || '';
-                    const buyPrice = Number(buyPriceStr);
-                    const sellPrice = Number(sellPriceStr);
-                    const permissionLevel = Number(permLevelStr);
-
-                    if (
-                        displayName &&
-                        minecraftId &&
-                        icon &&
-                        !isNaN(buyPrice) &&
-                        !isNaN(sellPrice) &&
-                        !isNaN(permissionLevel)
-                    ) {
-                        const result = shopAdminManager.updateShopItem(categoryName, null, selectedEntry.id, {
-                            buyPrice,
-                            sellPrice,
-                            permissionLevel,
-                            icon,
-                            minecraftId,
-                            displayName
-                        });
-                        player.sendMessage(result.message);
-                    } else {
-                        player.sendMessage('§4Invalid data. Please check all fields.');
-                    }
-                } else {
-                    // Delete
-                    const result = shopAdminManager.removeItem(categoryName, null, selectedEntry.id);
-                    player.sendMessage(result.message);
-                }
-            } else {
-                // subCategory
-                return showPanel(player, `shopAdminSubCategoryItemPanel_${selectedEntry.name}`, {
-                    ...context,
-                    subCategoryName: selectedEntry.name
-                });
-            }
-            return showPanel(player, panelId, { ...context, page: 1 });
-        }
-        // Handle pagination
-        let newPage = page;
-        const totalPages = Math.ceil(allEntries.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-        const buttonIndex = selection && selection > 3 ? selection - 4 - paginatedEntries.length : -1;
-
-        if (hasPrev && buttonIndex === 0) {
-            newPage--;
-        } else if (hasNext) {
-            newPage++;
-        }
-        return showPanel(player, panelId, { ...context, page: newPage });
-    }
-
-    if (panelId.startsWith('shopAdminCategoryActionPanel_')) {
-        const categoryName = panelId.replace('shopAdminCategoryActionPanel_', '');
-
-        if (selection === 0) {
-            // Edit
-            const shopConfig = getShopConfig();
-            const category = shopConfig.categories[categoryName];
-            const form = new ModalFormData()
-                .title('Edit Category')
-                .textField('Category Name', 'Enter new name', { defaultValue: categoryName })
-                .textField('Icon', 'Enter icon texture path', { defaultValue: category.icon });
-            const editCatResponse = await utils.uiWait(player, form);
-            if (editCatResponse.canceled) {
-                return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, context);
-            }
-            const values = (editCatResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const newName = values[0] as string;
-            const newIcon = values[1] as string;
-
-            if (newName) {
-                const result = shopAdminManager.editCategory(categoryName, newName, newIcon || '');
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, 'shopManagementPanel', { ...context, page: 1 });
-        }
-        if (selection === 1) {
-            // Delete
-            showConfirmationDialog(player, {
-                title: 'Confirm Deletion',
-                body: 'Are you sure?',
-                confirmButtonText: '§4Yes, Delete',
-                cancelButtonText: '§2No, Cancel',
-                onConfirm: () => {
-                    const result = shopAdminManager.deleteCategory(categoryName);
-                    player.sendMessage(result.message);
-                    return showPanel(player, 'shopManagementPanel', { ...context, page: 1 });
-                },
-                onCancel: () => {
-                    return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, context);
-                }
-            });
-            return;
-        }
-        if (selection === 2) {
-            // Back
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, context);
-        }
-    }
-
-    if (panelId.startsWith('shopAdminSubCategoryItemPanel_')) {
-        const { categoryName, subCategoryName, page = 1 } = context;
-        if (selection === 0) {
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, context);
-        }
-        if (selection === 1) {
-            // Add Item
-            return showPanel(player, `shopAddItemPanel_${categoryName}`, { ...context, subCategoryName });
-        }
-        if (selection === 2) {
-            // Edit Subcategory
-            return showPanel(player, `shopAdminSubCategoryActionPanel_${subCategoryName}`, context);
-        }
-
-        const shopConfig = getShopConfig();
-        const subCategory = shopConfig.categories[categoryName].subCategories[subCategoryName];
-        const items = Object.keys(subCategory.items).map((id) => ({ id, ...subCategory.items[id], type: 'item' }));
-        const paginatedItems = getPaginatedItems(items, page);
-        const selectedItem =
-            selection && selection > 2
-                ? (paginatedItems[selection - 3] as {
-                      id: string;
-                      displayName: string;
-                      icon: string;
-                      buyPrice: number;
-                      sellPrice: number;
-                      permissionLevel: number;
-                  })
-                : undefined;
-
-        if (selectedItem) {
-            const form = new ActionFormData()
-                .title('Edit Item')
-                .button('Edit', 'textures/ui/icon_setting')
-                .button('Delete', 'textures/ui/trash');
-            const itemActionResponse = await utils.uiWait(player, form);
-            if (itemActionResponse.canceled) {
-                return showPanel(player, panelId, context);
-            }
-            const itemActionSelection = (itemActionResponse as ActionFormResponse).selection;
-            if (itemActionSelection === 0) {
-                // Edit
-                const masterItem = allItems[selectedItem.id] || {};
-                const editForm = new ModalFormData()
-                    .title(`Edit Item: ${selectedItem.id}`)
-                    .textField('Display Name', 'e.g., Magical Sword', {
-                        defaultValue: selectedItem.displayName || masterItem.displayName
-                    })
-                    .textField('Minecraft Item ID', 'e.g., minecraft:diamond_sword', {
-                        defaultValue: masterItem.itemId
-                    })
-                    .textField('Icon Path', 'e.g., textures/items/diamond_sword', {
-                        defaultValue: selectedItem.icon || masterItem.icon
-                    })
-                    .textField('Buy Price', '-1 to disable', { defaultValue: String(selectedItem.buyPrice) })
-                    .textField('Sell Price', '-1 to disable', { defaultValue: String(selectedItem.sellPrice) })
-                    .textField('Permission Level', 'e.g., 1024', {
-                        defaultValue: String(selectedItem.permissionLevel)
-                    });
-
-                const editResponse = await utils.uiWait(player, editForm);
-                if (editResponse.canceled) {
-                    return showPanel(player, panelId, context);
-                }
-
-                const values = (editResponse as ModalFormResponse).formValues as string[];
-                const [displayName, minecraftId, iconStr, buyPriceStr, sellPriceStr, permLevelStr] = values;
-                const icon = iconStr || '';
-                const buyPrice = Number(buyPriceStr);
-                const sellPrice = Number(sellPriceStr);
-                const permissionLevel = Number(permLevelStr);
-
-                if (
-                    displayName &&
-                    minecraftId &&
-                    icon &&
-                    !isNaN(buyPrice) &&
-                    !isNaN(sellPrice) &&
-                    !isNaN(permissionLevel)
-                ) {
-                    const result = shopAdminManager.updateShopItem(categoryName, subCategoryName, selectedItem.id, {
-                        buyPrice,
-                        sellPrice,
-                        permissionLevel,
-                        icon,
-                        minecraftId,
-                        displayName
-                    });
-                    player.sendMessage(result.message);
-                } else {
-                    player.sendMessage('§4Invalid data. Please check all fields.');
-                }
-            } else {
-                // Delete
-                const result = shopAdminManager.removeItem(categoryName, subCategoryName, selectedItem.id);
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, panelId, { ...context, page: 1 });
-        }
-
-        // Handle pagination
-        let newPage = page;
-        const totalPages = Math.ceil(items.length / itemsPerPage);
-        const hasPrev = page > 1;
-        const hasNext = page < totalPages;
-        const buttonIndex = selection && selection > 2 ? selection - 3 - paginatedItems.length : -1;
-
-        if (hasPrev && buttonIndex === 0) {
-            newPage--;
-        } else if (hasNext) {
-            newPage++;
-        }
-        return showPanel(player, panelId, { ...context, page: newPage });
-    }
-
-    if (panelId.startsWith('shopAdminSubCategoryActionPanel_')) {
-        const subCategoryName = panelId.replace('shopAdminSubCategoryActionPanel_', '');
-        const { categoryName } = context;
-        if (selection === 0) {
-            // Edit
-            const shopConfig = getShopConfig();
-            const subCategory = shopConfig.categories[categoryName].subCategories[subCategoryName];
-            const form = new ModalFormData()
-                .title('Edit Subcategory')
-                .textField('Subcategory Name', 'Enter new name', { defaultValue: subCategoryName })
-                .textField('Icon', 'Enter icon texture path', { defaultValue: subCategory.icon });
-            const editSubCatResponse = await utils.uiWait(player, form);
-            if (editSubCatResponse.canceled) {
-                return showPanel(player, `shopAdminSubCategoryItemPanel_${subCategoryName}`, context);
-            }
-            const values = (editSubCatResponse as ModalFormResponse).formValues;
-            if (!values) return;
-            const newName = values[0] as string;
-            const newIcon = values[1] as string;
-
-            if (newName) {
-                const result = shopAdminManager.editSubCategory(categoryName, subCategoryName, newName, newIcon || '');
-                player.sendMessage(result.message);
-            }
-            return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, { ...context, page: 1 });
-        }
-        if (selection === 1) {
-            // Delete
-            showConfirmationDialog(player, {
-                title: 'Confirm Deletion',
-                body: 'Are you sure?',
-                confirmButtonText: '§4Yes, Delete',
-                cancelButtonText: '§2No, Cancel',
-                onConfirm: () => {
-                    const result = shopAdminManager.deleteSubCategory(categoryName, subCategoryName);
-                    player.sendMessage(result.message);
-                    return showPanel(player, `shopAdminCategoryPanel_${categoryName}`, { ...context, page: 1 });
-                },
-                onCancel: () => {
-                    return showPanel(player, `shopAdminSubCategoryActionPanel_${subCategoryName}`, context);
-                }
-            });
-            return;
-        }
-        if (selection === 2) {
-            // Back
-            return showPanel(player, `shopAdminSubCategoryItemPanel_${subCategoryName}`, context);
-        }
-    }
-
-    // ... (rest is same)
-    return;
 }
